docs/network-automation/examples/dashboard/requirements.txt
# Web Framework
flask>=2.0.0
werkzeug>=2.0.0

# Data Processing
pandas>=1.4.0
numpy>=1.21.0

# System Monitoring
psutil>=5.9.0

# HTTP Requests
requests>=2.31.0
urllib3>=1.26.0

# Environment Variables
python-dotenv>=1.0.0

# Data Visualization
plotly>=5.13.0

# Type Hints
typing-extensions>=4.7.1

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Development
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
